+++ src/molecules/Alert/Timer.tsx
@@ -1,0 +1,6 @@
+import { TimerStyled } from './styles';
+import { RdAlertTimerComponent } from './types';
+
+export const Timer: RdAlertTimerComponent = props => {
+    return <TimerStyled {...props} />;
+};